+++ src/main/java/com/legendsofvaleros/util/field/RangedValue.java
@@ -60,6 +60,6 @@ public class RangedValue {
 	}
 	
 	public long longValue() {
-		return (long)intValue();
+		return intValue();
 	}
 }
